refactor: variable value is not reassigned at parsing time instead
This occurs because at the moment to use a variable as value that previously has been reassigned dont apply the new value but the initial value of the variable.

The root of the problem is that, at parsing time, a parameter or any variable which his assigned value was a variable. the 'AssignmentExpression' (now called 'ReassignmentExpression') was not executed until the object dont reach the interpreter, this last part occurs after the variable value has been assigned