refactor: the super ultra mega refactor that will change ScrapLang forever
This refactor changes how source code files are read, as well as how expressions are parsed. As of this commit, the language is not interpreted — the entire source code is simply tokenized and converted into an AST. At the moment, not all expressions can be analyzed: only expressions that are a single value (such as numbers, strings, etc.) and others that only depend on binary operators. More complex expressions, such as those that use parentheses to denote operation precedence or that use unary operators, are not yet available.

A “parsing context” system is also yet to be implemented. This would allow a parser object to behave differently depending on what is being parsed at a given moment. For example, the comma (,) is used as an operator in expressions, but it has other uses in declarative contexts, such as in variable declarations — much like in C-derived languages.

Everything is subject to changes that will very likely happen!